+++ assets/js/dashboard/stats/modals/filter.js
@@ -18,7 +18,7 @@ export const FILTER_GROUPS = {
   'screen': ['screen'],
   'browser': ['browser', 'browser_version'],
   'os': ['os', 'os_version'],
-  'utm': ['utm_medium', 'utm_source', 'utm_campaign'],
+  'utm': ['utm_medium', 'utm_source', 'utm_campaign', 'utm_content', 'utm_term'],
   'entry_page': ['entry_page'],
   'exit_page': ['exit_page'],
   'goal': ['goal']